Added significant speed increase to point scrolling in the point list. BugID:725 Analysis: The renderer for the point buffer was using a slow brute force method of updating the points. It would remove all the old points from the renderer and then add the new updated points. This would force the renderer to do extra work adding and removing the same points it had before.
Resolution:
The renderer now only adds or removes points if the point list in the data object is longer or shorter than the previous list that was processed by the update method. Point properties
can change without having the point removed and then added back to the renderer. In this way selected points are only those points where the properties (color,opacity) are different
from the unselected points. The resulting improvement is an order of magnitude (10x) faster! :)

As a side note, since the point properties change the basic point must now keep track of the color it is supposed to be when it is unselected. A more elegant way to do this is to add a
property for the selected point and have a separate property object for the unselected point. Try to do this at a later date.

Reviewed by: