Added ability to add new tags to the 3d point buffer list. BugID:725 Analysis: The user needs to be able to add custom properties/tags to the 3d point buffer list.
Resolution:
Added a button and a line edit to the 3d point buffer GUI that will add a column to the point buffer.
The user can edit values in this column to add the property value for each point. The name is the column is what the user enters in the line edit. Duplicate names for column headers are
not accepted.

Reviewed by: