Save 3D point buffer to existing file Analysis: The "Save" function always opens a dialog to select a file, regardless of whether a file has already been chosen in a previous "Save" operation. This is disruptive to our experimental workflow which requires that the 3d window remain visible and that "Save" operations require as little interaction as possible.
If a file has already been selected, subsequent "Save"'s should
save to that file.

Resolution:
"Save" will now save to the same file as previously selected and only
present the user with a dialog on the first "Save" operation.

The automatic overwrite is not an expected behaviour for most users
but is important for our EP experimental workflow, so this should be
made optional. This has not been done.

A "Save As" operation should be implemented to allow the user to
change to a new file. This has not been done.
BugID: 725