Added "Save As" function for 3DPointBufferData object
Analysis:
The Save button would let user choose file with a dialog only first click,
then overwrite on every subsequent click.

Resolution:
Added a "Save As" option, which will open a dialog to let the user change
the file where the data will be saved at any time. The save button continues
to overwrite the previously selected file.
Current filename that is being saved to is now displayed at the bottom of the
form.

BugID: 725
Reviewed by: