Improved simulated trigger auto on/off feature
Analysis:
Was waiting for thread to establsih connection before updating simulated
triggers. Needed a better way and some user feedback on the connection.

Resolution:

Changed it so that connecting thread sends signal to slot in UI when connection
is established. Also added a new label to UI that tells user when/where
connection is on/off

BugID: 741
Reviewed by: