Using catheter SNR to change color not size
Analysis:

When checking the "use SNR" option, size of the spheres corresponding to coils
would change based on value of SNR. Wanted to change it to something less
distracting and more configurable.

Resolution:

Changed the color of the catheter spheres to traffic light colors (green for good,
yellow for middle and red for bad). Also added 2 spin boxes so that the values
of what SNRs are considered bad and good can be set in the controls

BugID: 745
Reviewed by: