Added camera movement in interaction mode to 3D volume objects
Analysis:

Had to switch between camera mode and interaction mode to move camera
after an interaction.

Resolution:

Added methods in the customQVTKwidget so that objects can hand off mouse
handling to camera control if we are not interacting with the object.

Implemented for 3D volumes. Plan to add 3D points and 2D planes

BugID: 754
Reviewed by: