Added volume catheter-following to the Cath Tracking plugin
Analysis:

Catheter following was not implemented

Resolution:

Added the ability to follow a catheter by setting the three slice planes of
a volume to intersect with the catheter automatically every time the position
updates.
Due to the design of the 3D volume render/data objects, the three slice planes
axial,sagittal and coronal could not be externally controlled, so several
signals and slots were added between the data and render objects to allow for
moving the planes.
The choice of using the plane normal or the original acquisition direction
was given, but as of now, only the normal direction moving is implemented

BugID: 679
Reviewed by: