Multiple Render Windows
Analysis:

Vurtigo could only switch between a single or split view of the same render
window. Interactions could only be done on the left side and only one
orientation widget existed for both views

Resolution:

Added the ability to add as many independant render windows as you want.
Each one can be interacted with independently.
Removing a render window "hides" it so that if the user clicks on "add render
window" the last one removed will be restored (with the view it had when we
removed it)

NOTE: Adding multiple render windows meant that every mouse event is now associated
with the window it came from. All the necessary changes to the mouse handlers
of all objects were made.

BugID: 766
Reviewed by: