Added point placement and landmark registration to Vurtigo
Analysis:

We want to be able to select points on volumes, planes and surfaces for
labeling as well as use as landmarks for registration.

Resolution:

Users can open the point placement dialog from the main menu to enable point
placement (using the right mouse button) in the 3D view.
Users can select/create 3D point objects where the points will be stored.
Points will be colored by a random color (chosen from a preset list)
Users can edit the label for each point, which will place a 3D label near the
point which will always face the camera (right now they always face the camera
of the first render window)

Users can open the registration dialog in order to perform landmark registration
using a volume and a set of source and landmark points.
Point placement in this dialog works as above, with several differences:
1. source points can only be placed on the source volume, same for target
(right-clicking anywhere else will do nothing)
2. The color of a point will match the corresponding color in the other set
(source/target) if it exists
3. Source points are labeled S0,S1,S2,... and Targets T0,T1,T2,...

Toggling the Sync wheel checkbox will apply any wheel scroll interactions done
on one volume to be mirrored on the other.

Once a set of source and target points have been chosen, the user can choose
which method of registration to perform:
1. Rigid - only translations and rotations
2. Similarity - Rigid + isotropic scaling
3. Affine - full 4x4 transform
4. Thin Plate Spline - non-linear transform

Performing registration will create 2 new objects, a new Volume and a new
3D Point object. Both will be named after the method of registration, the volume
and the IDs of the source and target points used.
The new volume will retain the plane transforms,visiblity,phase,... of the source.
The new points will be labeled by which method was selected.
A Mean-Squared-Error will be displayed at the bottom of the dialog (given by
the distance between the target and new points in global coordinates

NOTE: The creation of a set of transformed points is currently only implemented
for linear transforms.

FOR THIS COMMIT:
This is the first commit and not all features discussed above are added yet
- Registration does not yet create the new registered object
- no object sync yet
- no color matching yet
- no point labels yet
- no MSE measure yet

BugID: 764 765
Reviewed by: