Combined Camera/Interact/Placement modes into one + registration GUI improvements
Resolution:

Camera movement, object interaction and point placement can all be combined
without the need for seperate modes:
1. If no object is selected, all mouse events move the camera
2. Double-clicking will select an object and interactions work as they do
in interactions mode, but if we click or drag anywhere NOT on an interaction
widget, we revert back to camera movements.
3. right-clicking will place a point (but is only enabled when a suitable dialog
is open)

NOTE: If an object is selected, mouse wheel events will always scroll the object
regardless of where the cursor is (you must deselct the object to regain zoom)

Some GUI improvements to the landmark registration dialog.
The active object (source or target) is now highlighted with a raised frame so as
to make it easier on the user to distinguish between them

BugID: 767
Reviewed by: