Added object synchronization in landmarkRegistration dialog
Resolution:

Added a general framework for synching two render objects together (each object
maintains a list of objects it is synched to which can be used to perform the
same action on multiple objects at once.

This was implemented for 3D Volumes as passing on wheel events to all objects
synched with the volume.
This was used in the landmark registration dialog to allow the user to sync the
source and target volumes so that scrolling the slice plane in one volume also
scrolled the same plane in the other (allowing for easier placement of
matching landmarks)

NOTE: Actual registration code is still in progress, so button is disabled

BugID: 765
Reviewed by: