Methods of landmark registration now work
Resolution:

For linear transforms, found the transform by using the appropriate vtk landmark
transform, then created a new Volume and set it to that transform.
The target points were similarly transformed into a new set of points.

For Thin plate spline, first performed an affine transform to align the points,
then used the vtkthinplatespline transform to warp the volume's image based
on that transform (making sure to center the image using vtkImageChangeInformation
before and after the reslicing)

BugID: 765