Added auto-save feature in CathHistory + more user feedback
Analysis:

We wanted to be able to save the marked points directly to the file specified
by the 3D point object we were marking to.

Resolution:

Added an auto-save checkbox the user can use to have the marked points save not
only in the 3D point object but directly to disk as well.
If no filename was specified for the 3D point object to save to, we alert the user
and do not save until a filename has been provided.

BugID: 753
Reviewed by: