Fixed numLoc bug and made tracking exclusive to one catheter
Analysis:

When a catheter was chosen to be tracked, but the user would turn off all the
coils in it, Vurtigo would crash due to a divide by zero.
Also, planes and volumes could be set to track more than one catheter, which
we don't support yet (the latest one would be tracked but the GUI would not update)

Resolution:

Checked to see if numLoc was 0 before continuing.
Made it so that a user could not turn on tracking if the selected volume/plane
was already paired with a catheter

BugID: 679
Reviewed by: