Added a reset button and more GUI/user feedback changes
Analysis:

Want to be able to reset the points and start a new session from within the
CathHistory plugin.

Resolution:

Added a reset button that saves the current points, clears the point data and
resets the set to 0.
Also added some changes to the feedback (they disappear after some time) and
handles the cases where the user cancels the file dialog correctly.

Made a small change to the file dialog in 3Dpoint data object so that when a
user does a "save as" the dialog starts from the path of the previous file that
was saved (maybe make this default behaviour from now on?)

BugID: 753
Reviewed by: