Object renaming
Analysis:

We couldn't rename objects after they had been created

Resolution:

Added a "rename object" action to the object menu.

Also connected signals in point placement and registration so that when an obejct is renamed, it updates the combo boxes
Did this also for when objects are created/removed through the objectManager (this led to a couple of bug fixes from
accessing the NULL pointers of destroyed but not yet updated objects in the combo boxes)

NOTE: do we need to do this for every combo box in Vurtigo that doesn't already do it?

BugID: 771
Reviewed by: