Added Volume Coloring
Analysis:

Users could not change the colors of the slice planes that showed the volumes

Resolution:

Added buttons in the 3-plane options that let users choose the color of each individual plane.
The color chosen will also determine the color of the box outline of that plane (as well as color the border
around the button itself for quick reference).

BugID: 775
Reviewed by: