GPU Volume rendering
Analysis:

Vurtigo was doing 3D volume rendering on the CPU but could be slow on
slower systems. Wanted to get the speed up of GPU.

Resolution:

Added the option of GPU rendering. This only becomes an option if the
hardware/volume supports it.

NOTE: GPU rendering does not support all the CPU versions raycast functions
(MIP,isosurface,...), so they are disabled in GPU mode

BugID: 760
Reviewed by: