Save/Load EP Object
Analysis:

Want to be able to save/load the current state of an EP object

Resolution:

User can now save/load the current state of the EP object.
Vurtigo will save 2 xml files: the EP object (original contour points+properties)
and the 3D points (if we saved with color mapping on).
If the user tries to save the object with spacing smaller than 1.0, they will be
warned that load times could be long and are given the option to keep the spacing
information or not.

Properties saved:
- point opacity
- surface opacity
- in plane spacing (optional)
- cross plane spacing (optional)
- current Phase

If the 3D point xml cannot be found, the user will be warned and loading will
continue once the user has dismissed the warning

NOTE: a small change was made to the general loadObject in mainWindow so that
an individual file could be specified without the need to open a new file dialog.
Also, the ids of the loaded objects are returned.

NOTE: each saved file is approximately 4MB in size

BugID: 702
Reviewed by: