Change a selected volume/2D slice window/level using the mouse
Resolution:

If the user has a volume or 2D slice selected, then holding the shift key and then
clicking and dragging on the render window changes the window level.
X axis changes the level part
Y axis changes the window part

These changes are also made to the window/level dialog (if the dialog is open,
the bar graph version of the window/level will update as well)

The range of window/level values is taken into account so mouse movements are mapped
to roughly equal steps for different volumes/slices

NOTE: running out of modifiers. shift is reused because alt and ctrl are already
mapped to other things too

BugID: 761
Reviewed by: