Converted all object types to QString
Resolution:

Object types are now stored as QStrings (so that new types can be added and set
to whatever you want)
Backwards compatability was maintained as much as possible:
- Vurtigo can still load in objects saved using the old enum style
- Plugins can still request objects using the old enum style (it is converted to
string when stored and saved)

WARNING!

Because of the way QHash works, the order of traversal is arbitrary. This
means that the ordering of the objects in the object tree has changed because
we now hash by a string rather than the enum.

WARNING!!!
the basehandle method for plugins getObjectType() returns a QString now.
This can break plugins that used this method. Please fix your plugins!!!

BugID: 787
Reviewed by: