Saving volumes to disk
Resolution:

Vurtigo can now save Volumes to files. Saving a volume produces 2 files:
An XML file containing the header (used for loading), the image type (DICOM or image)
and the trigger list.
The other file is a VTK image file created using the XMLImageDataWriter.

NOTE: If the image file does not exist (user moved/renamed it) nothing will be
created, even if the header XML is loaded

TODO: Need to save volume transform as well

BugID: 702
Reviewed by: