Customizable plugin loading path fix
Analysis:
The plugin "libpath" in QSettings does not correctly customize the plugin
loading path. The application directory (the directory containing
the main executable) is used instead. The QLibrary::isLibrary check
that is being used does not check that the library exists, only that
the extension of the given filename is a valid suffix for a loadable
library.

In this case, the output "Plugin: could not be loaded" does not specify
which plugin couldn't be loaded; this is because the plugin loader sets
the name of the plugin to the empty string when it does not find the plugin.

Resolution:
Checking that the file exists in the application directory and given
directory before trying to load it from either of these locations.
This will allow the user to put their plugins in a location other
than the main executable path.

If the plugin loading fails, the original plugin name specified by
the user is output along with the plugin loader's resolved plugin path.

BugID: 769