New catheter coloring scheme
Analysis:

Catheter objects stored coordinates and SNR and could alter the color of all
points based on SNR values.

Resolution:

We wanted a more extensible framework for Vurtigo catheters:

Catheters now only store XYZ coordinates. Any other (double valued)
properties can be programmatically added (ex. SNR)

The option is now given to the user to choose a property (ex. SNR) and a Vurtigo
Color Transfer Function (which can be created programmatically, within Vurtigo
(using the New Object menu command), or editing a Vurtigo XML file).
The catheter points will then be rendered in a color based on the lookup table
value of that property.

Vurtigo color transfer functions now support specifying the midpoint and sharpness
at each color value.
The midpoint and sharpness values specified when adding a node are used to
control the transition to the next node. The midpoint is the normalized distance
between control points at which the curve reaches the median Y value.
A sharpness of 0 yields a piecewise linear function and a sharpness of 1 yields
a piecewise constant function.
Users can now change the color of each node in the function both by manually
setting the RGB values ([0,255]) or by double-clicking on the color in the
color column and choosing a new color using the color dialog

BugID: 745
Reviewed by: