New HelloWorld Plugin showcasing new Object FrameWork
Resolution:

Modified the existing HelloWorld Plugin so that it showed an example of how
to implement a new render/data object (with GUI) from within a Plugin and how
to register the new object type using the basehandler.

Also, set the default mouse event handlers in rtRenderObject to camera control,
so that if an object does not define any interaction, the camera will take over.

BugID: 788
Reviewed by: