Catheter Save/Load
Resolution:

Added the save/load methods for catheters.
Catheters will save all the spline properties.
If a color scheme was selected during the save, the color transfer function that
was chosen will be saved in a separate file. When loading, Vurtigo will attempt
to load in the saved color transfer function to return the color scheme to the
state the object was in when it was saved (a warning will appear if the file is
missing, but the object will still load, just without the color scheme)

BugID: 702
Reviewed by: