framerate bug workaround + new Permanent objects
Analysis:

Attempting to delete the frame rate label object would cause Vurtigo to crash

Resolution:

The Frame Rate label object has 2 issues that cause Vurtigo to crash:

1. Label objects are the only objects to crash Vurtigo when they are deleted and
the object tree attempts to remove its widget. (This could be because (not sure), similar to
colortransferfunc and piecewisefunc objects, the label objects set up their
widgets manually (no calls to setupUI) but do not explicitly remove all their
child widgets in the destructor)

2. The frame rate object itself is special in the sense that the time manager will
always expect a frame rate object to be present to continually update the text
to the latest calculated frame rate. A problem then appears if we delete the
frame rate object while the time manager is still trying to update the frame rate.

The workaround to this, is to make the frame rate object a "permanent" object in
Vurtigo.
By doing so, a general method for making objects permanent was added (Vurtigo and
plugins (through the basehandler) can make objects permanent, but nothing can
make them deletable again during runtime).

A permanent object cannot be deleted from the main menu (nothing will happen)
and the delete command will not appear in that object's context menu.

BugID: 784
Reviewed by: