Added the ability to customize the orientation marker widget. Fixed minor bugs relating to the orientation marker widget.
Analysis:

a) The orientation marker widget did not provide intuitive information about the coordinate system.
b) When dragging the orientation marker widget in interactive mode, the 3D viewing camera would also respond to mouse-move events.
c) Vurtigo could be made to crash by making the orientation marker widget interactive and then hidden, and then playing with it.
d) The vtkAxesActor does not calculate it's bounds correctly, leading to viewing problems when changing its orientation. Some parts of it would be clipped under certain transformations.

Resolution:

a) The user can now select between several default orientation widgets (Labeled Axes, Body, Head, Heart). These are specified in an XML document which can easily be extended. Models are in the 3ds format.
b) The orientation marker widget was not pickable from the perspective of a vtkPicker, so no prop was considered picked when the widget was being dragged. The orientation widget now sets itself as picked whenever the mouse interacts with it appropriately. This prevents the camera control from receiving events.
c) When set to hidden, the orientation marker widget did not disable itself. This meant that if it was previously set to interactive, and then disabled, it would still be considered interactive. When a user then clicked on it Vurtigo would segfault. To fix it all that was required was to set interactive to false before disabling it.
d) rtAxesActor subclasses vtkAxesActor and overrides its incorrect method GetBounds().

BugID: 782, 791
Reviewed by: