Gate catheter by trigger delay
Analysis:

In order to smooth the motion of the catheter, we want to  be able to specify
a bounds on the trigger delay so as to only update the catheter position during
a specific phase.

Resolution:

If gating is enabled, the global trigger delay is checked to see if it is within
the user specified min/max bounds. If not, we don't update the catheter position.

BugID: 797
Reviewed by: