Added an XYZ orientation marker prop.
Analysis:
The coordinate types included 'VTK' which behaved identically to 'HFS'.
Furthermore, since the relabeling of the vtkAxesActor was removed, there was no longer a way to have the labels XYZ.
Patient relative coordinates do not always make sense.

Resolution:
A special case rtAxesActor was added to represent the XYZ coordinate system.

BugID: 782
Reviewed by: