Fix for object visibility/# window misalignment
Analysis:

When a render window was removed, its components (render widget, renderer, etc.)
were being maintained and used as the next "new" window.
In contrast, each object was removing its visibility status from the render window
when it was removed.
When the object was then deleted, we removed the object from the render window
and then set the visibility in that window to false...a visibility that no longer
existed.
Hence the crash.

Resolution:

Since we do not differentiate between removing an object from a render window
on deletion or just turning visibility off, rather than creating a new method
just for deletion or adding a check to every removeFromRenderer object method,
I just added a check in the setVisible3D method, that ignored requests to set
the visibility on the extra render windows that we save for later "new" use
but are invisible to the object itself.

Also, removed a commented out line of code (doesn't really pertain to this
bug fix)

BugID: 800
Reviewed by: