Fixed Vurtigo freezing when interacting with 3D points during placement
Analysis:

When either the point placement or registration dialogs were open, if a 3D
point object was selected in the dialog and the user then tried to interact with
the object via its compass widget, Vurtigo would freeze.

Resolution:

The signal that connected changes to the 3D point object to the update of the
point tables in the dialogs was located in the same slot that it connected to.
This apparently caused some kind of infinite loop that hung Vurtigo.

The signal was moved to a seperate method.

BugID: 806
Reviewed by: